requirements.txt
torch==2.0.0
torchvision==0.15.0
streamlit==1.28.1
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.2
pillow==10.0.0

# ADD THESE (VERY IMPORTANT)
pandas==2.1.4
xarray==2023.7.0
plotly==5.18.0